+++ src/stores/index.ts
@@ -1,9 +1,13 @@
+import { useCommonStore } from './commonStore';
 import { useGameStore } from './gameStore';
 import { useLanguageStore } from './languageStore';
 import { useSettingStore } from './settingStore';
+import { useWalletStore } from './walletStore';
 
 export {
+    useCommonStore,
     useGameStore,
     useLanguageStore,
-    useSettingStore
+    useSettingStore,
+    useWalletStore
 };